+++ flask_inputfilter/Filter/Base64ImageDownscaleFilter.py
@@ -1,11 +1,10 @@
 from __future__ import annotations
 
 import base64
-import binascii
 import io
 from typing import Any, Optional
 
-from PIL import Image, UnidentifiedImageError
+from PIL import Image
 
 from flask_inputfilter.Filter import BaseFilter
 
@@ -39,13 +38,7 @@ class Base64ImageDownscaleFilter(BaseFilter):
             image = Image.open(io.BytesIO(base64.b64decode(value)))
             return self.resize_picture(image)
 
-        except (
-            binascii.Error,
-            UnidentifiedImageError,
-            OSError,
-            ValueError,
-            TypeError,
-        ):
+        except (OSError, ValueError, TypeError):
             return value
 
     def resize_picture(self, image: Image) -> str: