+++ flask_inputfilter/Model/__init__.py
@@ -1,2 +1,9 @@
+import shutil
+
 from .ExternalApiConfig import ExternalApiConfig
-from .FieldModel import FieldModel
+
+if shutil.which("g++") is not None:
+    from ._FieldModel import FieldModel
+
+else:
+    from .FieldModel import FieldModel